+++ src/Neo/Cryptography/Helper.cs
@@ -18,7 +18,6 @@ using Org.BouncyCastle.Crypto.Modes;
 using Org.BouncyCastle.Crypto.Parameters;
 using System;
 using System.Buffers.Binary;
-using System.IO.Hashing;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
@@ -32,7 +31,6 @@ namespace Neo.Cryptography
     /// </summary>
     public static class Helper
     {
-        private const int DefaultXxHash3Seed = 40343;
         private static readonly bool IsOSX = RuntimeInformation.IsOSPlatform(OSPlatform.OSX);
 
         /// <summary>
@@ -118,30 +116,6 @@ namespace Neo.Cryptography
         {
             using var sha256 = SHA256.Create();
             return sha256.ComputeHash(value);
-        }
-
-        /// <summary>
-        /// Computes the 32-bit hash value for the specified byte array using the xxhash3 algorithm.
-        /// </summary>
-        /// <param name="value">The input to compute the hash code for.</param>
-        /// <param name="seed">The seed used by the xxhash3 algorithm.</param>
-        /// <returns>The computed hash code.</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int XxHash3_32(this ReadOnlySpan<byte> value, long seed = DefaultXxHash3Seed)
-        {
-            return HashCode.Combine(XxHash3.HashToUInt64(value, seed));
-        }
-
-        /// <summary>
-        /// Computes the 32-bit hash value for the specified byte array using the xxhash3 algorithm.
-        /// </summary>
-        /// <param name="value">The input to compute the hash code for.</param>
-        /// <param name="seed">The seed used by the xxhash3 algorithm.</param>
-        /// <returns>The computed hash code.</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int XxHash3_32(this byte[] value, long seed = DefaultXxHash3Seed)
-        {
-            return XxHash3_32(value.AsSpan(), seed);
         }
 
         /// <summary>
