+++ tests/Neo.UnitTests/Ledger/UT_StorageKey.cs
@@ -12,6 +12,7 @@
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Cryptography;
+using Neo.Extensions;
 using Neo.SmartContract;
 using System;
